+++ changelog.md
@@ -6,4 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   Enable the `allowDeclareFields` option of `@babel/preset-typescript`.
+
 ## [1.0.0] - 2021-06-14